Released v0.0.9 Changed the 'prepublish' task to clear the /lib/ folder before building. Added a 'start' task which simply starts jest in watch mode.
Updated to the latest flow, which caught some more errors with the
definition of the 'dispatch' parameter in the config, the dispatch always
requires an Action now.

Also fixed a serious type error in the validators. The typings suggested
that they returned ValidationError's, but instead they return functions
which return ValidationError's.

Changed the reducer file so Actions are typed according to the recommendation
of flow. See: https://flow.org/en/docs/frameworks/redux/.

Finally updated the jest version to 20.