Updated to redux-form 7.3.0.
`redux-form` 7.3.0 changes the way the `validate` property on a `Field`
works, it now re-renders the component when the `validate` prop changes.
This has the effect of resetting the component each time `validate` changes.
When the component resets all errors disappear.

This means that for `jarb-redux-form` that we must now cache the
validator functions to prevent the re-render. If we do not cache the
validators we never see any errors! All validators are now calculated
once per `JarbField` and stored internally for reuse.

Also updated a lot of dev dependencies such as react, jest and enzyme.

Installed `flow-typed` to catch extra errors, and enabled flow
in the tests as well, which caught some new errors. Fixed the errors
by updating the models and fixing the type for the regex validators.
Also the test now use `spyOn` instead of `jest.spyOn` to get around
flow errors.

Decided that the package is ready for a 1.0.0 release. Therefore
adding a CHANGELOG.md file.